Fix serialization to account for sRGB flag when loading textures.
Also fixed the copy constructor for Plane and Cube so that it creates a new VAO, VBO and loads textures again.  This need to be done because the destructor was finally deleting the VAO/VBO and textures, so when the old copy died (a particular problem when the old copy was a local variable) the new copy was pointing to correct vertex data.